requirements.txt
gradio==4.44.1
openai>=1.54.0
mcp>=1.2.0
httpx>=0.27.0
python-dotenv>=1.0.1
pydantic>=2.7.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
